Move pipeline running tools to separate directory
