+++ pdrs4all/postprocess/extract_templates.py
